+++ code/APIServer/src/main/java/com/cc/pic/api/src/pojo/CategoryTag.java
@@ -24,7 +24,7 @@ public class CategoryTag extends Model<CategoryTag> {
     private Long categoryTagId;
     private Long pid;
     private Integer moduleType;
-    private Integer orderIndex;
+    private Long orderIndex;
     private String categoryName;
